pyproject.toml
[project]
name = "sm-ai-v2-back"
version = "0.1.0"
description = "Backend for SoundMind AI V2 - Multi-Modal Agentic System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "langchain>=0.1.10",
    "langchain-openai>=0.0.8",
    "langchain-anthropic>=0.1.4",
    "langgraph>=0.0.26",
    "mcp>=0.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "chromadb>=0.4.22",
    "tiktoken>=0.6.0",
    "langchain-community>=0.0.20",
    "pypdf>=4.0.0",
    "python-multipart>=0.0.9",
    "docx2txt>=0.8",
    "python-pptx>=0.6.21",
    "langgraph-supervisor>=0.0.1",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
